+++ cloud-operations/quota-monitoring/main.tf
@@ -62,6 +62,7 @@ module "cf" {
   bundle_config = {
     source_dir  = "cf"
     output_path = var.bundle_path
+    excludes    = null
   }
   # https://github.com/hashicorp/terraform-provider-archive/issues/40
   # https://issuetracker.google.com/issues/155215191