+++ networking/onprem-google-access-dns/versions.tf
@@ -13,5 +13,17 @@
 # limitations under the License.
 
 terraform {
-  required_version = ">= 0.12.6"
+  required_version = ">= 1.0.0"
+  required_providers {
+    google = {
+      source  = "hashicorp/google"
+      version = ">= 4.0.0"
+    }
+    google-beta = {
+      source  = "hashicorp/google-beta"
+      version = ">= 4.0.0"
+    }
+  }
 }
+
+