+++ networking/private-cloud-function-from-onprem/main.tf
@@ -187,7 +187,6 @@ module "test-vm" {
   }
   network_interfaces = [{
     addresses  = null
-    alias_ips  = null
     nat        = false
     network    = module.vpc-onprem.self_link
     subnetwork = module.vpc-onprem.subnet_self_links["${var.region}/${var.name}-onprem"]
@@ -208,6 +207,7 @@ module "function-hello" {
   bundle_config = {
     source_dir  = "${path.module}/assets"
     output_path = "bundle.zip"
+    excludes    = null
   }
   bucket_config = {
     location             = var.region