+++ tests/modules/cloud_run/test_plan.py
@@ -1,0 +1,50 @@
+# Copyright 2021 Google LLC
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#      http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+
+import os
+import pytest
+
+
+FIXTURES_DIR = os.path.join(os.path.dirname(__file__), 'fixture')
+
+
+@pytest.fixture
+def resources(plan_runner):
+  _, resources = plan_runner(FIXTURES_DIR)
+  return resources
+
+
+def test_resource_count(resources):
+  "Test number of resources created."
+  assert len(resources) == 5
+
+def test_iam(resources):
+  "Test IAM binding resources."
+  bindings = [r['values'] for r in resources if r['type']
+              == 'google_cloud_run_service_iam_binding']
+  assert len(bindings) == 1
+  assert bindings[0]['role'] == 'roles/run.invoker'
+
+def test_audit_log_triggers(resources):
+  "Test audit logs Eventarc trigger resources."
+  audit_log_triggers = [r['values'] for r in resources if r['type']
+              == 'google_eventarc_trigger' and r['name'] == 'audit_log_triggers']
+  assert len(audit_log_triggers) == 1
+
+def test_pubsub_triggers(resources):
+  "Test Pub/Sub Eventarc trigger resources."
+  pubsub_triggers = [r['values'] for r in resources if r['type']
+              == 'google_eventarc_trigger' and r['name'] == 'pubsub_triggers']
+  assert len(pubsub_triggers) == 2